CAAFI Data: Computer Aversion, Attitudes, and Familiarity Inventory

Study: This dataset has data collected on the computer aversion, attitudes, and familiarity inventory.

Usage

data(caafidata)

Format

A data frame with 794 rows and 30 variables.

q1

I enjoy using computers.

q2

Being able to use a computer is important to me.

q3

I keep up with the latest computer hardware.

q4

Computers are beneficial because they save people time.

q5

I like using word-processing programs.

q6

I feel like a fool when I am using a computer and others are around.

q7

I am smart enough to use a computer.

q8

I avoid using computers whenever possible.

q9

I do not understand how to use computer software (e.g., word-processing programs, spreadsheet programs, etc.).

q10

I feel that I understand how to use computer files, documents, and folders.

q11

I use a computer input device every day (e.g., a keyboard, a touch pad, a mouse).

q12

I can use a computer to successfully perform tasks.

q13

I can add new hardware to a computer.

q14

I enjoy reading computer magazines.

q15

When I use a computer, I am afraid that I will damage it.

q16

I enjoy connecting new computer accessories.

q17

I must have a reference manual or a help file to run computer software.

q18

E-mail is an easy way to communicate with people.

q19

I use e-mail every day.

q20

I am comfortable changing (installing/upgrading) computer software.

q21

I often read computer books.

q22

My friends often ask me computer-related questions.

q23

I often read computer magazines.

q24

Overall, I feel that I don't know how to use a computer.

q25

Computers are too scientific for me.

q26

When using a computer, I often lose data.

q27

I enjoy learning to use new software programs.

q28

I like to use computer input devices such as a keyboard, a touch pad, a mouse, etc.

q29

Using a computer is entertaining.

q30

I keep up with the latest computer software.

Details

The instructions were: Below is a list of items describing many of the thoughts and experiences that people have with computers. After reading each statement, circle the number that best describes how true or how false the statement is as it applies to you at this time. If you have no opinion about the item, circle ‘‘0”, but please use this option only if it is absolutely necessary. Be sure to circle only one number. Please do your best to respond to each item.

Scale -3 to 3: absolutely false, neutral, absolutely true

Computer Familiarity: Items 3, 13-14, 16, 20-23, 27, and 30.

Computer Attitudes: Items 1-2, 4-5, 8, 11, 18-19, and 28-29.

Computer Aversion: Items 6-7, 9-10, 12, 15, 17, and 24-26.