DASS Data: Depression, Anxiety, and Stress Inventory

Study: The DASS is a measurement scale that examines the depression, anxiety, and stress of an individual.

Usage

data(dassdata)

Format

A data frame with 794 rows and 30 variables.

Q1

I found it hard to wind down.

Q2

I was aware of dryness of my mouth.

Q3

I couldn't seem to experience any positive feeling at all.

Q4

I experienced breathing difficulty.

Q5

I found it difficult to work up the initiative to do things.

Q6

I tended to over-react to situations.

Q7

I experienced trembling (eg, in the hands).

Q8

I felt that I was using a lot of nervous energy.

Q9

I was worried about situations in which I might panic and make a fool of myself.

Q10

I felt that I had nothing to look forward to.

Q11

I found myself getting agitated.

Q12

I found it difficult to relax.

Q13

I felt down-hearted and blue.

Q14

I was intolerant of anything that kept me from getting on with what I was doing.

Q15

I felt I was close to panic.

Q16

I was unable to become enthusiastic about anything.

Q17

I felt I wasn't worth much as a person.

Q18

I felt that I was rather touchy.

Q19

I was aware of the action of my heart in the absence of physical exertion.

Q20

I felt scared without any good reason.

Q21

I felt that life was meaningless.

Details

The instructions were: Please read each statement and select a number 0, 1, 2 or 3 that indicates how much the statement applied to you over the past week. There are no right or wrong answers. Do not spend too much time on any statement, but please answer each question.

Scale 0-3: did not apply to me, applied to me to some degree, applied to me to a considerable degree, or applied to me very much

Depression: Questions 3, 5, 10, 13, 16, 17, 21

Anxiety: Questions 2, 4, 7, 9, 15, 19, 20

Stress: 1, 6, 8, 11, 12, 14, 18