Exploratory Factor Analysis Practice Dataset

Study: This dataset has data on the Openness to Experience scale collected as part of an undergraduate honor's thesis project.

Usage

data(efa)

Format

A data frame with 99 rows and 21 variables.

o1

Believe in the importance of art.

o2

Have a vivid imagination.

o3

Tend to vote for liberal political candidates.

o4

Carry the conversation to a higher level.

o5

Enjoy hearing new ideas.

o6

Enjoy thinking about things.

o7

Can say things beautifully.

o8

Enjoy wild flights of fantasy.

o9

Get excited by new ideas.

o10

Have a rich vocabulary.

o11

Am not interested in abstract ideas.

o12

Do not like art.

o13

Avoid philosophical discussions.

o14

Do not enjoy going to art museums.

o15

Tend to vote for conservative political candidates.

o16

Do not like poetry.

o17

Rarely look for a deeper meaning in things.

o18

Believe that too much tax money goes to support artists.

o19

Am not interested in theoretical discussions.

o20

Have difficulty understanding abstract ideas.

condition

a group condition each participant received

Details

The instructions were: Below are some phrases describing people's behaviors. Please use the rating scale below to describe how accurately each statement describes you. Describe yourself as you generally are now, not as you wish to be in the future. Describe yourself as you honestly see yourself in relation to other people of your gender and of roughly your same age. Please read each statement carefully, and then check the box that corresponds to your response.

Scale: very inaccurate, moderately inaccurate, neither inaccurate nor accurate, moderately accurate, very accurate