Meaning and Purpose in Life Data

Study: This data includes three measures of meaning and purpose in life for exploring latent variables or multi-trait multi-methods analysis.

Usage

data(meaningdata)

Format

A data frame with 567 rows and 50 variables.

p1

I am usually (completely bored to exuberant, enthusiastic)

p2

Life to me seems (completely routine to always exciting)

p3

In life I have (no goals or aims at all to very clear goals and aims)

p4

My personal existence is (utterlying meaningless without purpose to very purposeful and meaningful)

p5

Every day is (exactly the same to constantly new)

p6

If I could choose, I would (prefer never to have been born to like nine more lives just like this one)

p7

After retiring, I would (loaf completely the rest of my life to do some of the exciting things I have always wanted to do)

p8

In achieving life goals I have (made no progress whatever to progressed to complete fulfillment)

p9

My life is (empty, filled only with despair to running over with exciting good things)

p10

If I should die today, I would feel that my life had been (completely worthless to very worthwhile)

p11

In thinking of my life, I (often wonder why I exist to always see a reason for my being here)

p12

As I view the world in relation to my life, the world (completely confuses me to fits meaningfully with my life)

p13

I am a (very irresponsible person to very responsible person)

p14

Concerning man's freedom to make his own choices, I believe man is (completely bound by limitations of heridity and environment to absolutely free to make all life choices)

p15

With regard to death, I am (unprepared and afraid to prepared and unafraid)

p16

With regard to suicide, I have (thought of it seriously as a way out to never given it a second thought)

p17

I regard my ability to find a meaning, purpose, or mission in life as (practically none to very great)

p18

My life is (out of my hands and controlled by external factors to in my hands and I am in control of it)

p19

Facing my daily tasks is (a painful and boring experience to a source of pleasure and satisfaction)

p20

I have discovered (no mission or purpose in life to clear-cut goals and a satisfying life purpose)

m1

I understand my life's meaning.

m2

I am lookin gof rsomething ath makes my life feel meaningful.

m3

I am always looking to find my life's ppurpose.

m4

My life has a clear sense of purpose.

m5

I have a good sense of what makes my life meaningful.

m6

I have discovered a satifying life purpose.

m7

I am lalways searching for something that makes my life feel significant.

m8

I am seeking a purpose or mission for my life.

m9

My life has no clear purpose.

m10

I am searching for meaning in my life.

s1

I think about the ultimate meaning of life.

s2

I have experienced the feeling that while I am destined to accomplish something important, I cannot quite put myfinger on just what it is.

s3

I try new activities or areas of interest, and then these soon lose their attractiveness.

s4

I feel that some element which I can't quite define is missing fro mmy life.

s5

I am restless.

s6

I feel that the greatest fulfillment of my life lies wyet in the future.

s7

I hope for something exciting in the future.

s8

I daydream of finding a new place for my life and a new identity.

s9

I feel the lack of – and a need to find – a real meaning and purpose in my life.

s10

I think of achieving something new and different.

s11

I seem to change my main objective in life.

s12

The mystery of life puzzles and disturbs me.

s13

I feel myself in need of a 'new lease on life'.

s14

Before I achieve one goal, I start out toward a different one.

s15

I feel the need for adventure and 'new worlds to conquer'.

s16

Over my lifetime I have felt a strong urge to find myself.

s17

On occasion I have thought that I had found what I was looking for in life, only to have it vanish later.

s18

I have been aware of all-powerful and consuming purpose toward which my life has been directed.

s19

I have sensed a lack of worthwhile job to do in life.

s20

I have felt a determination to achieve something f ar beyond the ordinary.

Details

The Meaning in Life Questionnaire is scored from 1 absolutely untrue to 7 absolutely true. These items are marked with a M in the dataset. The Purpose in Life Questionnaire is scaled from 1 to 7 varying by the question and is marked P in the dataset. Last, the Seeking of Neotic Goals scale ranges from 1 never to 7 constantly and is marked with S in the dataset.